% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalModule.R
\name{calculatingExploreData}
\alias{calculatingExploreData}
\title{Calculate what is needed to explore the model with the data in the R environment}
\usage{
calculatingExploreData(results)
}
\arguments{
\item{results}{the results that need to be caluclated}
}
\description{
Calculate what is needed to explore the model with the data in the R environment
}
\author{
Matthew L. Fidler
}
