% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createShinyPKModel.R
\name{createShinyPKModel}
\alias{createShinyPKModel}
\title{This function creates a nlmixr2 model by specifying the absorption, distribution and elimination}
\usage{
createShinyPKModel(
  absorption = c("IV/Infusion/Bolus", "First Order", "Transit"),
  distribution = c("1 compartment", "2 compartment", "3 compartment"),
  elimination = c("linear", "Michaelis-Menten"),
  ntransit = 3
)
}
\arguments{
\item{absorption}{This specifies the absorption method and can be:
- IV/Infusion/Bolus -- no changes to the base model
- First Order -- adds a depot compartment
- Transit -- adds a transit compartment of size `ntransit`}

\item{distribution}{number of compartments in the model;  Can be:
- 1 compartment
- 2 compartment
- 3 compartment}

\item{elimination}{the type of elimination of PK.  Currently supports:
- linear
- Michaelis-Menten}

\item{ntransit}{number of transit compartments (by default 3)}

\item{linCmt}{boolean, is this going to use a linear compartment solution}
}
\value{
a PK model according to the specifications above;  This will return a list of string
}
\description{
In general this is used in shiny to create the base PK model
}
\examples{

# By defualt this creates a one comparmtent ODE model from the `nlmixr2lib` model library
createShinyPKModel()

createShinyPKModel(absorption="First Order")
}
