% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalModule.R
\name{nlmixr2model}
\alias{nlmixr2model}
\title{The nlmixr2model function}
\usage{
nlmixr2model(app = FALSE)
}
\value{
A Shiny UI object.
}
\description{
The main UI function for the nlmixr2Shiny app using Shiny (no miniUI).
}
