% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalModule.R
\name{nlmixr2shiny}
\alias{nlmixr2shiny}
\title{The nlmixr2shiny function}
\usage{
nlmixr2shiny(app = FALSE)
}
\arguments{
\item{app}{Should this be run as an app? If TRUE, it will run as a
standard Shiny app. If FALSE, it will run as a Shiny gadget.}
}
\value{
A Shiny UI object.
}
\description{
The main UI function for the nlmixr2Shiny app using Shiny (no miniUI).
}
