% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{ini}
\alias{model}
\alias{lotri}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{lotri}{\code{\link[lotri]{lotri}}}

  \item{rxode2}{\code{\link[rxode2]{ini}}, \code{\link[rxode2]{model}}}
}}

