% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxUiResidDf.R
\name{residDf}
\alias{residDf}
\alias{residDf.rxUi}
\alias{residDf.norm}
\alias{residDf.t}
\alias{residDf.cauchy}
\alias{residDf.dnorm}
\alias{residDf.default}
\title{This creates the residual data frame for}
\usage{
residDf(line)

\method{residDf}{rxUi}(line)

\method{residDf}{norm}(line)

\method{residDf}{t}(line)

\method{residDf}{cauchy}(line)

\method{residDf}{dnorm}(line)

\method{residDf}{default}(line)
}
\arguments{
\item{line}{line to parse}
}
\value{
error lines for the model
}
\description{
This creates the residual data frame for
}
\author{
Matthew L. Fidler
}
\keyword{internal}
