% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxUiResidDf.R
\name{residInfo}
\alias{residInfo}
\alias{residInfo.rxUi}
\alias{residInfo.norm}
\alias{residInfo.t}
\alias{residInfo.cauchy}
\alias{residInfo.dnorm}
\alias{residInfo.default}
\title{This creates the residual data frame for}
\usage{
residInfo(line)

\method{residInfo}{rxUi}(line)

\method{residInfo}{norm}(line)

\method{residInfo}{t}(line)

\method{residInfo}{cauchy}(line)

\method{residInfo}{dnorm}(line)

\method{residInfo}{default}(line)
}
\arguments{
\item{line}{line to parse}
}
\value{
error lines for the model
}
\description{
This creates the residual data frame for
}
\author{
Matthew L. Fidler
}
\keyword{internal}
